requirements.txt
# Voice Guard v4.0 - Anti-Clawback Backend
# 핵심 종속성 (AWS RDS + B2 WORM + Whisper)
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai-whisper==20240930
boto3==1.35.76
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.18
pydantic==2.10.3
# Phase 2: 카카오 알림톡 (Solapi REST API v4)
requests==2.32.3
# Phase 4: Notion 미러 동기화 워커
httpx==0.28.1
# Phase 1: Care Plan/Billing Excel 파싱
pandas>=2.2.0
openpyxl>=3.1.0
